Reorganización del codigo para adaptarse a arquitectura hexagonal
