--- src/main/java/com/acl/municipalidad/domain/exceptions/EmailAlreadyExistsException.java
@@ -1,7 +1,0 @@
-package com.acl.municipalidad.domain.exceptions;
-
-public class EmailAlreadyExistsException extends RuntimeException {
-    public EmailAlreadyExistsException(String message) {
-        super(message);
-    }
-}
