--- src/main/java/com/acl/municipalidad/domain/exceptions/EmailNotFoundException.java
@@ -1,7 +1,0 @@
-package com.acl.municipalidad.domain.exceptions;
-
-public class EmailNotFoundException extends RuntimeException{
-    public EmailNotFoundException(String message) {
-        super(message);
-    }
-}
