--- src/main/java/com/acl/municipalidad/domain/response/AuthenticationRequest.java
@@ -1,15 +1,0 @@
-package com.acl.municipalidad.domain.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AuthenticationRequest {
-    private String email;
-    private String password;
-}
