--- src/main/java/com/acl/municipalidad/domain/response/RegisterRequest.java
@@ -1,16 +1,0 @@
-package com.acl.municipalidad.domain.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class RegisterRequest {
-    private String username;
-    private String email;
-    private String password;
-}
