--- src/main/java/com/acl/municipalidad/infrastructure/database/mysql/controllers/AuthController.java
@@ -1,54 +1,0 @@
-package com.acl.municipalidad.infrastructure.database.mysql.controllers;
-
-import com.acl.municipalidad.application.AuthService;
-import com.acl.municipalidad.domain.exceptions.EmailAlreadyExistsException;
-import com.acl.municipalidad.domain.exceptions.EmailNotFoundException;
-import com.acl.municipalidad.domain.response.AuthResponse;
-import com.acl.municipalidad.domain.response.AuthenticationRequest;
-import com.acl.municipalidad.domain.response.RegisterRequest;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@Slf4j
-@RequiredArgsConstructor
-@RequestMapping("/api/v1/auth")
-//TODO:Cambiar el CORS
-@CrossOrigin(origins = "*")
-public class AuthController {
-
-    private final AuthService authService;
-
-    @PostMapping("/register")
-    public ResponseEntity<AuthResponse> register(@RequestBody RegisterRequest request) {
-        try {
-            AuthResponse response = authService.register(request);
-            return ResponseEntity.ok(response);
-        } catch (EmailAlreadyExistsException e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(AuthResponse.builder().error(e.getMessage()).build());
-        }
-    }
-
-    @PostMapping("/authenticate")
-    public ResponseEntity<AuthResponse> authenticate(@RequestBody AuthenticationRequest request){
-        try {
-            AuthResponse response = authService.authenticate(request);
-            return ResponseEntity.ok(response);
-        } catch (EmailNotFoundException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(
-                    AuthResponse.builder()
-                            .error(e.getMessage())
-                            .build());
-        }
-        catch (AuthenticationException e) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(
-                    AuthResponse.builder()
-                            .error("Credenciales incorrectas.")
-                            .build());
-        }
-    }
-}
