--- src/main/java/com/acl/municipalidad/infrastructure/database/mysql/entity/Role.java
@@ -1,6 +1,0 @@
-package com.acl.municipalidad.infrastructure.database.mysql.entity;
-
-public enum Role {
-    USER, ADMIN
-}
-
