+++ src/main/java/com/acl/municipalidad/security/JwtService.java
@@ -1,0 +1,86 @@
+package com.acl.municipalidad.security;
+
+import com.acl.municipalidad.user.infrastructure.entity.UserEntity;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import io.jsonwebtoken.SignatureAlgorithm;
+import io.jsonwebtoken.io.Decoders;
+import io.jsonwebtoken.security.Keys;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.stereotype.Service;
+
+import java.security.Key;
+import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.function.Function;
+
+@Service
+public class JwtService {
+    @Value("${secretKey}")
+    private String secretKey;
+
+    public String generateToken(UserDetails userDetails){
+        Map<String, Object> extraClaims = new HashMap<>();
+        String role = userDetails.getAuthorities().stream()
+                .findFirst()
+                .map(GrantedAuthority::getAuthority)
+                .orElse("USER");
+        extraClaims.put("role", role);
+        if (userDetails instanceof UserEntity) {
+            extraClaims.put("email", ((UserEntity) userDetails).getEmail());
+        }
+        return generateToken(extraClaims, userDetails);
+    }
+
+    public String generateToken(Map<String,Object> extraClaims, UserDetails userDetails){
+        return Jwts.builder()
+                .setClaims(extraClaims)
+                .setIssuedAt(new Date(System.currentTimeMillis()))
+                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60 * 24))
+                .signWith(getSignInKey(), SignatureAlgorithm.HS256)
+                .compact();
+    }
+
+    public String getUserName(String token){
+        return getClaim(token, Claims::getSubject);
+    }
+
+    public String getEmail(String token) {
+        return getClaim(token, claims -> claims.get("email", String.class));
+    }
+
+    public <T> T getClaim(String token, Function<Claims, T> claimsResolver){
+        final Claims claims = getAllClaims(token);
+        return claimsResolver.apply(claims);
+    }
+
+    private Claims getAllClaims(String token) {
+        return Jwts
+                .parserBuilder()
+                .setSigningKey(getSignInKey())
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    private Key getSignInKey(){
+        byte[] keyBytes = Decoders.BASE64.decode(secretKey);
+        return Keys.hmacShaKeyFor(keyBytes);
+    }
+
+    public boolean validateToken(String token, UserDetails userDetails) {
+        final String username = getUserName(token);
+        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
+    }
+
+    private boolean isTokenExpired(String token) {
+        return getExpiration(token).before(new Date());
+    }
+
+    private Date getExpiration(String token) {
+        return getClaim(token, Claims::getExpiration);
+    }
+}
