+++ src/main/java/com/acl/municipalidad/user/domain/model/Role.java
@@ -1,0 +1,5 @@
+package com.acl.municipalidad.user.domain.model;
+
+public enum Role {
+    USER, ADMIN
+}
