+++ src/main/java/com/acl/municipalidad/user/domain/model/User.java
@@ -1,0 +1,13 @@
+package com.acl.municipalidad.user.domain.model;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class User {
+    private Long id;
+    private String email;
+    private String password;
+    private Role role;
+}
