+++ src/main/java/com/acl/municipalidad/user/domain/port/AuthServicePort.java
@@ -1,0 +1,7 @@
+package com.acl.municipalidad.user.domain.port;
+
+import com.acl.municipalidad.user.infrastructure.adapter.request.AuthenticationRequest;
+
+public interface AuthServicePort {
+    String authenticateUser(AuthenticationRequest request);
+}
