+++ src/main/java/com/acl/municipalidad/user/infrastructure/adapter/request/AuthenticationRequest.java
@@ -1,0 +1,17 @@
+package com.acl.municipalidad.user.infrastructure.adapter.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+
+@Data
+public class AuthenticationRequest {
+    @NotEmpty(message = "Email is required")
+    @Email(message = "Email should be valid")
+    private String email;
+
+    @NotEmpty(message = "Password is required")
+    @Size(min = 6, message = "Password must be at least 6 characters long")
+    private String password;
+}
