+++ src/main/java/com/acl/municipalidad/user/infrastructure/adapter/response/ApiResponse.java
@@ -1,0 +1,11 @@
+package com.acl.municipalidad.user.infrastructure.adapter.response;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@AllArgsConstructor
+@Data
+public class ApiResponse {
+    private String message;
+    private Object data;
+}
