+++ src/main/java/com/acl/municipalidad/user/infrastructure/entity/Role.java
@@ -1,0 +1,6 @@
+package com.acl.municipalidad.user.infrastructure.entity;
+
+public enum Role {
+    USER, ADMIN
+}
+
