+++ src/main/java/com/acl/municipalidad/user/infrastructure/exceptions/EmailAlreadyExistsException.java
@@ -1,0 +1,7 @@
+package com.acl.municipalidad.user.infrastructure.exceptions;
+
+public class EmailAlreadyExistsException extends RuntimeException {
+    public EmailAlreadyExistsException(String message) {
+        super(message);
+    }
+}
