+++ src/main/java/com/acl/municipalidad/user/infrastructure/exceptions/EmailNotFoundException.java
@@ -1,0 +1,7 @@
+package com.acl.municipalidad.user.infrastructure.exceptions;
+
+public class EmailNotFoundException extends RuntimeException{
+    public EmailNotFoundException(String message) {
+        super(message);
+    }
+}
