+++ src/main/java/com/acl/municipalidad/user/infrastructure/repository/UserJpaRepository.java
@@ -1,0 +1,10 @@
+package com.acl.municipalidad.user.infrastructure.repository;
+
+import com.acl.municipalidad.user.infrastructure.entity.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserJpaRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByEmail(String email);
+}
